Refactor validate slug into mixin
