+++ app/utils/channel-const.js
@@ -2,4 +2,4 @@ export default {
 	titleMinLength: 4,
 	titleMaxLength: 32,
 	descriptionMaxLength: 280
-};
+}
